+++ packages/simple_graphql_client_gen/lib/graphql_post.dart
@@ -9,15 +9,28 @@ final class GraphqlResponse {
   const GraphqlResponse(this.data, this.errors);
 
   final JsonObject? data;
+  final GraphqlErrors? errors;
+}
+
+@immutable
+final class GraphqlErrors implements Exception {
+  const GraphqlErrors(this.errors);
+
+  /// nonEmpty
   final IList<GraphqlError> errors;
 }
 
 @immutable
 final class GraphqlError implements Exception {
-  const GraphqlError(this.message, this.code);
+  const GraphqlError({
+    required this.message,
+    required this.path,
+    required this.extensionsCode,
+  });
 
   final String message;
-  final String? code;
+  final IList<String> path;
+  final String? extensionsCode;
 }
 
 Future<GraphqlResponse> graphQLPost({
@@ -44,19 +57,29 @@ Future<GraphqlResponse> graphQLPost({
     utf8.decode(response.bodyBytes),
   );
 
+  final errors = jsonValue
+          .getObjectValueOrNull('errors')
+          ?.getAsArrayWithDecoder((errorJson) {
+        final message =
+            errorJson.getObjectValueOrNull('message')?.asStringOrNull();
+        final path = IList(errorJson
+            .getObjectValueOrNull('path')
+            ?.getAsArray()
+            ?.map((element) => element.asStringOrThrow()));
+        final extensionsCode = errorJson
+            .getObjectValueOrNull('extensions')
+            ?.getObjectValueOrNull('code')
+            ?.asStringOrNull();
+        return GraphqlError(
+          message: message ?? 'Unknown error',
+          path: path,
+          extensionsCode: extensionsCode,
+        );
+      }) ??
+      const IListConst([]);
+
   return GraphqlResponse(
     jsonValue.getObjectValueOrThrow('data').asJsonObjectOrNull(),
-    jsonValue
-            .getObjectValueOrNull('errors')
-            ?.getAsArrayWithDecoder((errorJson) {
-          final message =
-              errorJson.getObjectValueOrNull('message')?.asStringOrNull();
-          final code = errorJson.getObjectValueOrNull('code')?.asStringOrNull();
-          if (message == null) {
-            return null;
-          }
-          return GraphqlError(message, code);
-        }) ??
-        const IListConst([]),
+    errors.isEmpty ? null : GraphqlErrors(errors),
   );
 }